Consolidate `Startup` extensions into `WebHookMvcCoreBuilderExtensions`
